SG-master/requirements.txt
absl-py==0.7.1
astor==0.8.0
beautifulsoup4==4.8.0
boto==2.49.0
boto3==1.9.215
botocore==1.12.215
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Django==2.2.4
djangorestframework==3.10.2
docutils==0.15.2
elasticsearch==7.0.4
gast==0.2.2
gensim==3.8.0
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
jmespath==0.9.4
JPype1==0.7.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
konlpy==0.5.1
lxml==4.4.1
Markdown==3.1.1
matplotlib==3.1.1
nltk==3.4.5
numpy==1.17.0
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
smart-open==1.8.4
soupsieve==1.9.3
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
